Stp/Rstp: Ieee8021dInterfaceData: change priority to uint16_t
Stp/Rstp members are already uint16_t

+++ src/inet/linklayer/ieee8021d/common/Ieee8021dInterfaceData.h
@@ -30,7 +30,7 @@ class INET_API Ieee8021dInterfaceData : public InterfaceProtocolData
         /* The following values have same meaning in both STP and RSTP.
          * See Ieee8021dBDPU for more info.
          */
-        unsigned int priority;
+        uint16_t priority;
         unsigned int linkCost;
         bool edge;
